+++ changelog.md
@@ -5,3 +5,7 @@
 + added ding sounds to 'sell offer filled' and 'buy order filled'
 + fixed 'buy order set' cost per values being parsed incorrectly
 + changed gameload message
++ doesn't show 'not matched -- bz' error in chat
+
+# v1.2
++ added new messages to format